--- js/script.js
@@ -1,3 +1,0 @@
-import {photoDescriptionMocks} from './main.js';
-
-photoDescriptionMocks;
